Added fix for creating new custom messages.
When a new custom message was created the whole cache was reloaded because a new id was given by the database. This is now fixed. We'll get the next increment number from the database before putting the object in the database and add the id to the in cache stored object.